% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdc.R
\name{gdc_projects}
\alias{gdc_projects}
\title{List GDC projects}
\usage{
gdc_projects(size = 20, from = 0)
}
\arguments{
\item{size}{Number of results (default 20, max 10000)}

\item{from}{Offset for pagination (default 0)}
}
\value{
tibble: id, name, project_id, primary_site, disease_type,
program_name, summary_case_count, summary_file_count
}
\description{
List GDC projects
}
